Implement API Extension management
The new module gives the ability to
manage API Extensions.

XmlHelper abstract class is added because the
vcdApiClient can not parse the xml
response from the server

Some commponents are used in more
then one module, so they are moved
in one common module.

Translations for the new module are
added.

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>